Taschenrechner noch nicht fertig, bitte nicht anschauen

+++ src/aufgaben/AdditionsRechner.java
@@ -5,13 +5,13 @@ import java.util.Scanner;
 public class AdditionsRechner {
 
 public static void main(String[] args) {
-int zahl1,zahl2,ergebnis;
+long zahl1,zahl2,ergebnis;
 Scanner scanner = new Scanner(System.in);
 //
 System.out.println("Geben Sie bitte die erste Zahl ein:");
-zahl1=scanner.nextInt();
+zahl1=scanner.nextLong();
 System.out.println("Geben Sie bitte die zweite Zahl ein:");
-zahl2=scanner.nextInt();
+zahl2=scanner.nextLong();
 ergebnis=zahl1+zahl2;
 System.out.println("Das Ergebnis ist: " + ergebnis);
 }}
